CategoricalArray fixes
Fix mixing CategoricalArray with other kinds of arrays, which failed
due to problems with how NamedArrays handles CategoricalArray dimnames:
now we always unwrap categorical values when creating dimension names.

Fix the type of the dimnames vector when only CategoricalArrays are passed:
before this it was always Union{T, Missing}. Change it to always use the same eltype
as the input, which is consistent with other AbstractArrays.